% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{add}
\alias{add}
\title{add}
\usage{
add(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
expr ,  e.g., highlights a + b from a + b
}
\description{
This is a function that allows the user to highlight the mistakes,
in particular an extra element in a list
}
