% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{add_left}
\alias{add_left}
\title{add_left}
\usage{
add_left(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
expr  e.g., highlights a + from a + b
}
\description{
This is a function that allows the user to highlight the mistakes,
in particular the redundancies in the left-hand side of the expression.
}
