% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{add_right}
\alias{add_right}
\title{add_right}
\usage{
add_right(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
expr , e.g., highlights + b from a + b
}
\description{
This is a function that allows the user to highlight the mistakes,
in particular the redundancies in the right-hand side of the expression.
}
