% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{cal}
\alias{cal}
\title{Calligraphic font}
\usage{
cal(x)
}
\arguments{
\item{x}{an R symbol. This function is used to render the content in calligraphic font
in MathJax. In MathML, script font is used.}
}
\value{
The function cal is a wrapper for the identity function.
}
\description{
Calligraphic font
}
\examples{
mathjax(quote(K \%in\% cal(K)))

}
\seealso{
\code{\link[=identity]{identity()}}
}
