% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{decorations}
\alias{decorations}
\alias{roof}
\alias{boxed}
\alias{cancel}
\alias{phantom}
\alias{prime}
\alias{tilde}
\alias{over}
\alias{under}
\alias{underover}
\alias{hyph}
\alias{color}
\title{Identity functions for different decorations}
\usage{
roof(x)

boxed(x)

cancel(x)

phantom(x)

prime(x)

tilde(x)

over(x)

under(x)

underover(x)

hyph(x)

color(x)
}
\arguments{
\item{x}{the expression to render}
}
\value{
x
}
\description{
Identity functions for different decorations
}
\examples{
roof(1) + mean(2) + boxed(3) + cancel(4) + phantom(5) + prime(6) + tilde(7)

mathml(quote(roof(b) + mean(X) + boxed(3) + cancel(4) + phantom(5)))
}
