% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{denote}
\alias{denote}
\title{denote
This is a function that allows the user to insert abbreviations in the formula,
explain them and make the needed computations}
\usage{
denote(abbr, expr, info)
}
\arguments{
\item{abbr}{Abbreviation used in the text to refer to the calculation, for example 's_p' for the pooled
variance.}

\item{expr}{Expression: calculations to be made in order to obtain the value to which the abbreviation
refers to.}

\item{info}{Information: Explanation of the formula used to provide the value of the abbreviation.
e.g. 'the pooled variance'}
}
\value{
expr
e.g., x denotes a^2 + b
}
\description{
denote
This is a function that allows the user to insert abbreviations in the formula,
explain them and make the needed computations
}
