% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{dfrac}
\alias{dfrac}
\title{Division displayed as large fraction}
\usage{
dfrac(e1, e2)
}
\arguments{
\item{e1}{numerator}

\item{e2}{denominator}
}
\value{
e1 / e2
}
\description{
Division displayed as large fraction
}
\seealso{
\code{\link[=frac]{frac()}}, \code{\link[=over]{over()}}
}
