% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{dot}
\alias{dot}
\alias{nodot}
\alias{times}
\title{Multiplication}
\usage{
dot(e1, e2)

nodot(e1, e2)

times(e1, e2)
}
\arguments{
\item{e1}{numerator}

\item{e2}{denominator}
}
\value{
e1 * e2
}
\description{
Multiplication
}
