% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%.\%}
\alias{\%.\%}
\title{Product x * y, shown as x dot y}
\usage{
x \%.\% y
}
\arguments{
\item{x}{first factor}

\item{y}{second factor}
}
\value{
x * y
}
\description{
Product x * y, shown as x dot y
}
