% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%dbldown\%}
\alias{\%dbldown\%}
\title{Down double arrow, displayed as x dArr y}
\usage{
x \%dbldown\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
NA, it produces a downward double arrow
}
\description{
Down double arrow, displayed as x dArr y
}
