% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%dblup\%}
\alias{\%dblup\%}
\title{x uArr y}
\usage{
x \%dblup\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
x=y
it produces a upward double arrow
}
\description{
x uArr y
}
