% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%==\%}
\alias{\%==\%}
\title{Equivalence, shown as x == y}
\usage{
x \%==\% y
}
\arguments{
\item{x}{first argument}

\item{y}{second argument}
}
\value{
x == y
}
\description{
Equivalence, shown as x == y
}
