% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%=~\%}
\alias{\%=~\%}
\title{Congruence, shown as x =~ y}
\usage{
x \%=~\% y
}
\arguments{
\item{x}{first argument}

\item{y}{second argument}
}
\value{
x == y, e.g., a cong b
}
\description{
Congruence, shown as x =~ y
}
