% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%->\%}
\alias{\%->\%}
\title{Right arrow, presented as x -> y}
\usage{
x \%->\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
NA, it produces a right arrow
}
\description{
Right arrow, presented as x -> y
}
