% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%<=>\%}
\alias{\%<=>\%}
\title{x iff y}
\usage{
x \%<=>\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
x=y , it produces a double arrow double-sided
}
\description{
x iff y
}
