% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%+-\%}
\alias{\%+-\%}
\title{Plus Minus, it shows x and calculates x +- y}
\usage{
x \%+-\% y
}
\arguments{
\item{x}{first term}

\item{y}{second term}
}
\value{
c(x - y, x + y)
x plus min y
}
\description{
Plus Minus, it shows x and calculates x +- y
}
