% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%prop\%}
\alias{\%prop\%}
\title{Proportional, shown as x prop y}
\usage{
x \%prop\% y
}
\arguments{
\item{x}{first argument}

\item{y}{second argument}
}
\value{
x=y e.g, x prop y
}
\description{
Proportional, shown as x prop y
}
