% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%<->\%}
\alias{\%<->\%}
\title{x leftrightarrow y}
\usage{
x \%<->\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
x=y
it produces a doublrsides arrow
}
\description{
x leftrightarrow y
}
