% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%~~\%}
\alias{\%~~\%}
\title{Approximate equality, shown as x ~~ y}
\usage{
x \%~~\% y
}
\arguments{
\item{x}{first argument}

\item{y}{second argument}
}
\value{
The result of isTRUE(all.equal(x, y))
}
\description{
Approximate equality, shown as x ~~ y
}
