% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%up\%}
\alias{\%up\%}
\title{Up arrow, presented as x up y}
\usage{
x \%up\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
NA, it produces an upward arrow
}
\description{
Up arrow, presented as x up y
}
