% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{hook}
\alias{hook}
\title{Hook for custom symbols}
\usage{
hook(term = quote(t0), display = quote(subscript(t, 0)))
}
\arguments{
\item{term}{an R call or symbol/number. This is the expression to replace.}

\item{display}{an R call or symbol/number. This is shown instead of \emph{term}.}
}
\value{
TRUE on success
}
\description{
Hook for custom symbols
}
\examples{
hook(term=quote(t0), display=quote(subscript(t, 0)))
mathml(quote(t0))

}
