% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{instead}
\alias{instead}
\title{instead}
\usage{
instead(inst, of)
}
\arguments{
\item{inst}{instead}

\item{of}{of}
}
\value{
inst , e.g. a + c instead of a + b
}
\description{
This is a function that allows the user to highlight the mistakes,
in particular adds a curly bracket under the wrong term and it provides the
correct solutions.
}
