% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{mathjax}
\alias{mathjax}
\title{Mathjax output}
\usage{
mathjax(term = quote((a + b)^2L == a^2L + 2L * a * b + b^2L))
}
\arguments{
\item{term}{an R call or symbol/number. This function translates \emph{term} into a
LaTeX/MathJax string.}
}
\value{
A string with the MathJax representation or \emph{term}.
}
\description{
Mathjax output
}
\examples{
mathjax(term=quote((a + b)^2L == a^2L + 2L*a*b + b^2L))

}
\seealso{
\code{\link[=mathml]{mathml()}}
}
