% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{mathml_preproc}
\alias{mathml_preproc}
\title{Map R operators to their respective Prolog counterparts}
\usage{
mathml_preproc(query = quote(5\%\%2))
}
\arguments{
\item{query}{an R call or symbol/number. This function translates components of \emph{query}
into their respective counterparts from Prolog}
}
\value{
The translated query
}
\description{
Map R operators to their respective Prolog counterparts
}
\examples{
mathml_preproc(quote(5 \%\% 2))

}
\seealso{
\code{\link[=mathjax]{mathjax()}}, \code{\link[=mathml]{mathml()}}
}
