% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{name}
\alias{name}
\title{Add a name attribute to an element (most often, an R function)}
\usage{
name(x, name)
}
\arguments{
\item{x}{an R object, e.g., an R function}

\item{name}{the name of the object/function}
}
\value{
The object with the name attribute
}
\description{
Add a name attribute to an element (most often, an R function)
}
\examples{
f <- function(x) {sin(x)}
mathjax(call("integrate", name(f, "sin"), 0L, 2L*pi))

}
