% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{omit}
\alias{omit}
\title{omit}
\usage{
omit(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
NULL
e.g., remove a + b from a + b
}
\description{
This is a function that allows the user to highlight the mistakes,
in particular the omission of an element from a list.
}
