% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{omit_left}
\alias{omit_left}
\title{omit_left
This is a function that allows the user to highlight the mistakes,
in particular the omissions in the left-hand side of the expression}
\usage{
omit_left(expr)
}
\arguments{
\item{expr}{The expression, e.g. a + b}
}
\value{
substitute(expr)[[3]], e.g., b from a + b
}
\description{
omit_left
This is a function that allows the user to highlight the mistakes,
in particular the omissions in the left-hand side of the expression
}
