% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{omit_right}
\alias{omit_right}
\title{omit_right
This is a function that allows the user to highlight the mistakes,
in particular the omissions in the right-hand side of the expression}
\usage{
omit_right(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
substitute(expr)[[2]], e.g., a from a + b
}
\description{
omit_right
This is a function that allows the user to highlight the mistakes,
in particular the omissions in the right-hand side of the expression
}
