% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{over}
\alias{over}
\title{Sum over a range, e.g., sum x_i for i=1 to N}
\usage{
over(index = quote(i), from = 1, to = quote(N), fun = sum(1:10))
}
\arguments{
\item{index}{an R symbol, e.g., i.}

\item{from}{an R symbol, e.g., 1.}

\item{to}{an R symbol, e.g., N.}

\item{fun}{an R call. This is the return value of the function.}
}
\value{
The function over is a wrapper for the identity function, returning \emph{fun}
}
\description{
Sum over a range, e.g., sum x_i for i=1 to N
}
\examples{
mathjax(quote(over(index=i, from=1, to=N, fun=sum(1:10))))

}
\seealso{
\code{\link[=identity]{identity()}}
}
