% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{subscript}
\alias{subscript}
\title{Subscript. On the R side, this function is a wrapper of identity, but allows
for decorations.}
\usage{
subscript(fun = quote(x), sub = quote(i))
}
\arguments{
\item{fun}{an R call or symbol, e.g. sum(x). This is the return value of the function.}

\item{sub}{an R symbol or call, e.g., i}
}
\value{
The function over is a wrapper for the identity function, returning \emph{fun}
}
\description{
Subscript. On the R side, this function is a wrapper of identity, but allows
for decorations.
}
\examples{
mathjax(quote(subscript(sub=i, fun=x)))

}
\seealso{
\code{\link[=identity]{identity()}}
}
