% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{sum_over}
\alias{sum_over}
\title{sum over a range. On the R side, this function just returns the first
argument, but allows for decorations.}
\usage{
sum_over(x, from, to)
}
\arguments{
\item{x}{the object to be summed}

\item{from}{decoration for sum_from^to x_i}

\item{to}{decoration for sum_from^to x_i}
}
\value{
The function returns sum(x)
}
\description{
sum over a range. On the R side, this function just returns the first
argument, but allows for decorations.
}
\examples{
mathjax(quote(sum_over(x[i], i=1L, N)))

}
\seealso{
\code{\link[=sum]{sum()}}, \code{\link[=prod_over]{prod_over()}}
}
