% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{superscript}
\alias{superscript}
\title{Superscript. This is a wrapper for the identity function, but decorates the
result with a superscript.}
\usage{
superscript(fun = quote(A), sup = "*")
}
\arguments{
\item{fun}{an R call or symbol, e.g. x. This is the return value of the function.}

\item{sup}{an R symbol, e.g., "*"}
}
\value{
The function over is a wrapper for the identity function, returning \emph{fun}
}
\description{
Superscript. This is a wrapper for the identity function, but decorates the
result with a superscript.
}
\examples{
mathjax(quote(superscript(fun=A, sup="*")))

}
\seealso{
\code{\link[=identity]{identity()}}
}
