% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%dblup\%}
\alias{\%dblup\%}
\title{Up double arrow, displayed as x uArr y}
\usage{
x \%dblup\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
NA, it produces a upward double arrow
}
\description{
Up double arrow, displayed as x uArr y
}
