% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%down\%}
\alias{\%down\%}
\title{Down arrow, presented as x downarrow y}
\usage{
x \%down\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
NA, it produces a downward arrow
}
\description{
Down arrow, presented as x downarrow y
}
