% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%<=>\%}
\alias{\%<=>\%}
\title{If and only if condition, displayed as x <=> y}
\usage{
x \%<=>\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
NA, it produces a double arrow double-sided
}
\description{
If and only if condition, displayed as x <=> y
}
