% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%prop\%}
\alias{\%prop\%}
\title{Proportional, shown as x o< y}
\usage{
x \%prop\% y
}
\arguments{
\item{x}{first argument}

\item{y}{second argument}
}
\value{
NA
}
\description{
Proportional, shown as x o< y
}
