% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{\%<->\%}
\alias{\%<->\%}
\title{Double sided arrow, presented as x <-> y}
\usage{
x \%<->\% y
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
NA, it produces a double sided arrow
}
\description{
Double sided arrow, presented as x <-> y
}
