% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{instead}
\alias{instead}
\title{instead}
\usage{
instead(inst, of)
}
\arguments{
\item{inst}{the wrong term}

\item{of}{the correct term}
}
\value{
inst
}
\description{
This is a function that allows the user to highlight the mistakes,
in particular adds a curly bracket under the wrong term and it provides the
correct solutions.
}
\examples{
1 + instead(2, 3)

mathml(term=quote(1 + instead(2, 3)))

}
