% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathml.R
\name{mathml}
\alias{mathml}
\title{MathML output}
\usage{
mathml(
  term = quote((a + b)^2L == a^2L + 2L * a * b + b^2L),
  flags = NULL,
  env = globalenv()
)
}
\arguments{
\item{term}{an R call or symbol/number. This function translates \emph{term} into a
MathML string.}

\item{flags}{(default NULL)
list of flags that control the translation}

\item{env}{(default globalenv())
The R environment in which r_eval is being executed (see vignette for
details, "Ringing back to R").}
}
\value{
A string with the MathML representation or \emph{term}.
}
\description{
MathML output
}
\examples{
mathml(term=quote((a + b)^2L == a^2L + 2L*a*b + b^2L))

}
\seealso{
\code{\link[=mathjax]{mathjax()}}
}
